+++ packages/react/src/components/OneFilterPicker/OneFilterPicker.tsx
@@ -7,7 +7,7 @@ import { PresetsDefinition } from "./types"
 
 import { useEventEmitter } from "@/experimental/OneDataCollection/useEventEmitter"
 import { cn } from "@/lib/utils"
-import type { FiltersDefinition, FiltersState } from "./types"
+import type { FiltersDefinition, FiltersMode, FiltersState } from "./types"
 
 /**
  * Props for the Filters component.
@@ -26,6 +26,8 @@ export type OneFilterPickerRootProps<Definition extends FiltersDefinition> = {
   onChange: (value: FiltersState<Definition>) => void
   /** The children of the component */
   children?: React.ReactNode
+  /** The mode of the component */
+  mode?: FiltersMode
 }
 
 /**
@@ -93,6 +95,7 @@ const FiltersRoot = <Definition extends FiltersDefinition>({
   value,
   children,
   presetsLoading = false,
+  mode = "default",
   ...props
 }: OneFilterPickerRootProps<Definition>) => {
   const defaultFilters = useRef(value)
@@ -126,6 +129,7 @@ const FiltersRoot = <Definition extends FiltersDefinition>({
     <FiltersContext.Provider
       value={{
         ...props,
+        mode,
         presets: props.presets as PresetsDefinition<FiltersDefinition>,
         presetsLoading,
         value: localFiltersValue,
@@ -157,6 +161,7 @@ const FiltersControls = () => {
     setFiltersValue,
     presets,
     emitFilterChange,
+    mode,
   } = useContext(FiltersContext)
 
   const shownFilters = filters
@@ -181,6 +186,7 @@ const FiltersControls = () => {
         onOpenChange={setIsFiltersOpen}
         isOpen={isFiltersOpen}
         hideLabel={!!presets}
+        mode={mode}
       />
       {!!presets?.length && (
         <div className="flex items-center">
@@ -277,7 +283,7 @@ const OneFilterPicker = <Definition extends FiltersDefinition>(
           </div>
         )}
       </div>
-      <FiltersChipsList />
+      {(!props.mode || props.mode === "default") && <FiltersChipsList />}
     </FiltersRoot>
   )
 }