+++ packages/react/src/components/OneFilterPicker/components/FiltersControls.tsx
@@ -3,10 +3,12 @@ import { ButtonInternal } from "@/components/F0Button/internal"
 import { Filter } from "@/icons/app"
 import { useI18n } from "@/lib/providers/i18n"
 import { Popover, PopoverContent, PopoverTrigger } from "@/ui/popover"
+import { AnimatePresence, motion } from "motion/react"
 import { useEffect, useId, useMemo, useState } from "react"
+import { ArrowLeft } from "../../../icons/app"
 import { getFilterType } from "../filterTypes"
 import { FilterTypeContext, FilterTypeSchema } from "../filterTypes/types"
-import type { FiltersDefinition, FiltersState } from "../types"
+import type { FiltersDefinition, FiltersMode, FiltersState } from "../types"
 import { FilterContent } from "./FilterContent"
 import { FilterList } from "./FilterList"
 
@@ -17,6 +19,7 @@ interface FiltersControlsProps<Filters extends FiltersDefinition> {
   isOpen?: boolean
   onOpenChange?: (open: boolean) => void
   hideLabel?: boolean
+  mode?: FiltersMode
 }
 
 const DEFAULT_FORM_HEIGHT = 388
@@ -28,6 +31,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
   isOpen: controlledIsOpen,
   onOpenChange: controlledOnOpenChange,
   hideLabel,
+  mode = "default",
 }: FiltersControlsProps<Filters>) {
   const [selectedFilterKey, setSelectedFilterKey] = useState<
     keyof Filters | null
@@ -55,6 +59,19 @@ export function FiltersControls<Filters extends FiltersDefinition>({
     onOpenChange(false)
   }
 
+  const handleGoBack = () => {
+    if (selectedFilterKey) {
+      setSelectedFilterKey(null)
+    } else {
+      onChange(localFiltersValue)
+      onOpenChange(false)
+    }
+  }
+
+  const handleApplyFiltersSelection = () => {
+    handleGoBack()
+  }
+
   useEffect(() => {
     const getFirstFilterNotEmpty = () => {
       return Object.entries(localFiltersValue).find(([key, value]) => {
@@ -70,7 +87,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
       })
     }
 
-    if (isOpen) {
+    if (isOpen && mode === "default") {
       const firstFilterWithValue = getFirstFilterNotEmpty()
       if (firstFilterWithValue) {
         setSelectedFilterKey(firstFilterWithValue[0] as keyof Filters)
@@ -92,6 +109,94 @@ export function FiltersControls<Filters extends FiltersDefinition>({
     return maxHeight
   }, [filters])
   const id = useId()
+
+  if (mode === "compact") {
+    return (
+      <div className="flex items-center gap-2">
+        <ButtonInternal
+          variant="outline"
+          label={i18n.filters.label}
+          icon={Filter}
+          pressed={isOpen}
+          onClick={() => onOpenChange(!isOpen)}
+          aria-controls={isOpen ? id : undefined}
+          hideLabel
+          compact
+        />
+        <AnimatePresence mode="popLayout" propagate={false}>
+          {isOpen && (
+            <motion.div
+              initial={{ opacity: 0 }}
+              animate={{ opacity: 1 }}
+              exit={{ opacity: 0 }}
+              transition={{ duration: 0.1 }}
+              className="absolute bottom-0 left-0 right-0 top-0 z-20 bg-f1-background"
+            >
+              <div className="flex h-full flex-col transition-all">
+                <div className="pl-1.5 pt-1.5">
+                  <F0Button
+                    label="Back"
+                    icon={ArrowLeft}
+                    variant="ghost"
+                    size="sm"
+                    onClick={handleGoBack}
+                  />
+                </div>
+                <div className="flex flex-1">
+                  {selectedFilterKey ? (
+                    <motion.div
+                      key="filter-content"
+                      initial={{ opacity: 0 }}
+                      animate={{ opacity: 1 }}
+                      exit={{ opacity: 0 }}
+                      transition={{ duration: 0.2 }}
+                      className="h-full w-full bg-f1-background"
+                    >
+                      <FilterContent
+                        selectedFilterKey={selectedFilterKey}
+                        definition={filters}
+                        tempFilters={localFiltersValue}
+                        onFilterChange={updateFilterValue}
+                        isCompactMode
+                      />
+                    </motion.div>
+                  ) : (
+                    <motion.div
+                      key="filter-list"
+                      initial={{ opacity: 0 }}
+                      animate={{ opacity: 1 }}
+                      exit={{ opacity: 0 }}
+                      transition={{ duration: 0.2 }}
+                      className="h-full w-full bg-f1-background"
+                    >
+                      <FilterList
+                        definition={filters}
+                        tempFilters={localFiltersValue}
+                        selectedFilterKey={selectedFilterKey}
+                        onFilterSelect={(key: keyof Filters) =>
+                          setSelectedFilterKey(key)
+                        }
+                        onClickApplyFilters={handleApplyFilters}
+                        isCompactMode
+                      />
+                    </motion.div>
+                  )}
+                </div>
+                {selectedFilterKey && (
+                  <div className="sticky bottom-0 left-0 right-0 z-30 flex items-center justify-end gap-2 border border-solid border-transparent border-t-f1-border-secondary bg-f1-background p-2">
+                    <F0Button
+                      onClick={handleApplyFiltersSelection}
+                      label={i18n.filters.applySelection}
+                    />
+                  </div>
+                )}
+              </div>
+            </motion.div>
+          )}
+        </AnimatePresence>
+      </div>
+    )
+  }
 
   return (
     <div className="flex items-center gap-2">
@@ -127,6 +232,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
                 onFilterSelect={(key: keyof Filters) =>
                   setSelectedFilterKey(key)
                 }
+                onClickApplyFilters={handleApplyFilters}
               />
               {selectedFilterKey && (
                 <FilterContent
@@ -137,7 +243,7 @@ export function FiltersControls<Filters extends FiltersDefinition>({
                 />
               )}
             </div>
-            <div className="flex items-center justify-end gap-2 border border-solid border-transparent border-t-f1-border-secondary p-2">
+            <div className="flex items-center justify-end gap-2 border border-solid border-transparent border-t-f1-border-secondary bg-f1-background p-2">
               <F0Button
                 onClick={handleApplyFilters}
                 label={i18n.filters.applyFilters}