+++ packages/react/src/components/avatars/F0Avatar/F0Avatar.tsx
@@ -1,6 +1,7 @@
 import { ComponentProps, ReactNode } from "react"
 import { F0AvatarCompany, F0AvatarCompanyProps } from "../F0AvatarCompany"
 import { F0AvatarFile, F0AvatarFileProps } from "../F0AvatarFile"
+import { F0AvatarFlag, F0AvatarFlagProps } from "../F0AvatarFlag"
 import { F0AvatarPerson, F0AvatarPersonProps } from "../F0AvatarPerson"
 import { F0AvatarTeam, F0AvatarTeamProps } from "../F0AvatarTeam"
 import { AvatarSize } from "../internal/BaseAvatar"
@@ -15,6 +16,7 @@ export type AvatarVariant =
   | ({ type: "team" } & Omit<F0AvatarTeamProps, "size">)
   | ({ type: "company" } & Omit<F0AvatarCompanyProps, "size">)
   | ({ type: "file" } & Omit<F0AvatarFileProps, "size">)
+  | ({ type: "flag" } & Omit<F0AvatarFlagProps, "size">)
 
 export const F0Avatar = ({ avatar, size = "xs" }: AvatarProps): ReactNode => {
   switch (avatar.type) {
@@ -56,6 +58,16 @@ export const F0Avatar = ({ avatar, size = "xs" }: AvatarProps): ReactNode => {
       return (
         <F0AvatarFile
           file={avatar.file}
+          size={size as ComponentProps<typeof F0AvatarFile>["size"]}
+          badge={avatar.badge}
+          aria-label={avatar["aria-label"]}
+          aria-labelledby={avatar["aria-labelledby"]}
+        />
+      )
+    case "flag":
+      return (
+        <F0AvatarFlag
+          flag={avatar.flag}
           size={size as ComponentProps<typeof F0AvatarFile>["size"]}
           badge={avatar.badge}
           aria-label={avatar["aria-label"]}