+++ packages/react/src/hooks/datasource/useData.ts
@@ -250,6 +250,7 @@ export function useData<
     currentGrouping,
     grouping,
     idProvider = defaultIdProvider,
+    itemPreFilter,
   } = source
 
   const cleanup = useRef<(() => void) | undefined>()
@@ -263,7 +264,30 @@ export function useData<
     setError,
   } = useDataFetchState<R>()
 
+  const [filteredItemsCount, setFilteredItemsCount] = useState<number>(0)
+
   const { paginationInfo, setPaginationInfo } = usePaginationState()
+
+  useEffect(() => {
+    if (itemPreFilter) {
+      setRawData((currentData) => {
+        const originalItemsCount = currentData.length
+        const filteredData = currentData.filter(itemPreFilter)
+        const newItemsCount = filteredData.length
+        const filteredItemsCount = originalItemsCount - newItemsCount
+        setFilteredItemsCount(filteredItemsCount)
+        setPaginationInfo((info) =>
+          info
+            ? {
+                ...info,
+                total: info.total - filteredItemsCount,
+              }
+            : null
+        )
+        return filteredData
+      })
+    }
+  }, [itemPreFilter, setRawData, setPaginationInfo])
 
   // We need to use a ref to get the latest paginationInfo value
   // because the paginationInfo is updated asynchronously
@@ -323,7 +347,11 @@ export function useData<
   }
 
   const handleFetchSuccess = useCallback(
-    (result: PaginatedResponse<R> | SimpleResult<R>, appendMode: boolean) => {
+    (
+      result: PaginatedResponse<R> | SimpleResult<R>,
+      appendMode: boolean,
+      isLoadingYet?: boolean
+    ) => {
       /**
        * Call to the onResponse callback
        */
@@ -389,7 +417,7 @@ export function useData<
       )
       setError(null)
       setIsInitialLoading(false)
-      setIsLoading(false)
+      setIsLoading(!!isLoadingYet)
       setIsLoadingMore(false)
       isLoadingMoreRef.current = false
     },
@@ -596,12 +624,12 @@ export function useData<
 
         const subscription = observable.subscribe({
           next: (state) => {
-            if (state.loading) {
+            if (state.data) {
+              handleFetchSuccess(state.data, appendMode, state.loading)
+            } else if (state.loading) {
               setIsLoading(true)
             } else if (state.error) {
               handleFetchError(state.error)
-            } else if (state.data) {
-              handleFetchSuccess(state.data, appendMode)
             }
           },
           error: handleFetchError,
@@ -735,6 +763,8 @@ export function useData<
     }
   }, [])
 
+  const total = totalItems ? totalItems - filteredItemsCount : 0
+
   return {
     data,
     isInitialLoading,
@@ -745,7 +775,7 @@ export function useData<
     setPage,
     loadMore,
     mergedFilters,
-    totalItems,
+    totalItems: total,
   }
 }